Protect against multiple requests for same data.
Return 503 if another scoreboard requests production of the same videos as videos that are currently being produced.

Return 403 if a request from an external client (not localhost) attempts to produce videos.

We're still vulnerable to multiple clients running on localhost request overlapping videos at the same time, but that's unlikely. And anyway, we're going to restructure this so that arendad gets its direction from the chatroom. It just needs to listen for goals, and then publish the availability of replays. No problem! But not now!